Update README.md
    Executive Summary

In this section, address the following:

    Both questions of interest your group is seeking to answer.

    Motivation for both questions of interest.

    How does the analysis your group carried out in previous milestones answer these two questions of interest?

This section should be written with no technical jargon, and can be read and under-stood by any person with a college degree but has not taken any statistics classes. Think of how newspaper articles summarize results from a statistical study. As an example, look at this article from the New York Times (paragraphs 10 and 11) (https://www.nytimes.com/ 2018/08/13/well/an-underappreciated-key-to-college-success-sleep.html). This section should not be more than 2 pages.

    Data Description

Provide the following information of your data set:

a. Who or what are the data about.

    The source of the data set.

    Descriptions of the variables the reader will encounter later in the report. Clearly state which variables were the response variables. If you had to create your own variables, please clearly state that these variables were created and not part of the original data set, and describe how these were created.